+++ pymaid/fetch/__init__.py
@@ -56,7 +56,7 @@ from .. import core, utils, config, cache
 from navis import in_volume
 from .landmarks import get_landmarks, get_landmark_groups
 from .skeletons import get_skeleton_ids
-from .annotations import get_annotation_graph
+from .annotations import get_annotation_graph, get_entity_graph
 
 
 __all__ = ['get_annotation_details', 'get_annotation_id',
@@ -64,7 +64,8 @@ __all__ = ['get_annotation_details', 'get_annotation_id',
                   'get_arbor',
                   'get_connector_details', 'get_connectors',
                   'get_connector_tags',
-                  'get_contributor_statistics', 'get_edges', 'get_history',
+                  'get_contributor_statistics', 'get_edges', 'get_entity_graph',
+                  'get_history',
                   'get_logs', 'get_names', 'get_neuron',
                   'get_neurons', 'get_neurons_in_bbox',
                   'get_neurons_in_volume', 'get_node_tags', 'get_node_details',