+++ docs/source/whats_new.rst
@@ -12,7 +12,8 @@ What's new?
      -
    * - Next
      - In progress
-     - - signature of :func:`pymaid.get_annotation_graph` changed, allowing it to represent all semantic entities.
+     - - :func:`pymaid.get_annotation_graph` deprecated in favour of the new
+         :func:`pymaid.get_entity_graph`.
    * - 2.1.0
      - 04/04/22
      - With this release we mainly follow some renamed functions in ``navis`` but